Brings the repository-wide production-readiness pass onto dev: the workspace
now resolves from a clean clone, the lint backlog is cleared, and several
silently-broken APIs are fixed.

What this contributes:

- build: every internal dependency carries a version alongside its path and
  routes through the workspace table, all crates unify on 0.2.0, zenwave
  resolves from crates.io, and Cargo.lock is committed. aither-ort drops
  cuda/coreml from its default features; aither-llama and aither-ort move to
  current upstream releases.
- wasm32: aither-core, aither-openai, aither-gemini and aither-claude build
  for wasm32-unknown-unknown, with file:// attachment readers gated off a
  target that has no filesystem, and a CI job that actually checks it.
- core: a tool definition whose schema came off the wire no longer panics --
  ToolDefinition::from_parts returns InvalidSchema, and an MCP server
  offering an unusable schema has that one tool skipped rather than taking
  the process down. Tools::get now downcasts the tool instead of the box, so
  it stops always returning None. Registration rejects a tool with a blank
  description. Structured output carries a typed GenerateError.
- audio: the AudioGenerator and AudioTranscriber traits gained an error
  channel instead of swallowing failures behind a debug-only assert.
- acp: the ACP server is implemented -- session/new and session/prompt run
  the agent and stream real updates, replacing the echo placeholder.
- workspace: 23 unused dependencies dropped, oversized functions split into
  named stages, and the pedantic clippy backlog cleared across every crate.

Where the two sides disagreed, dev's newer work was kept and the branch's
fixes were ported onto it -- notably Event::ToolCallDelta, which is now
emitted from the restructured Claude stream state.

+++ cli/src/hook.rs
@@ -126,7 +126,7 @@ fn parse_terminal_args(arguments: &str) -> Option<(String, Option<String>)> {
 struct TerminalOutput {
     stdout: Option<String>,
     stderr: Option<String>,
-    exit_code: i32,
+    exit_code: i64,
     task_id: Option<String>,
     status: Option<String>,
 }
